disentangling-vae-master/requirements.txt
imageio==2.15.0
Imath==0.0.2
lpips==0.1.4
matplotlib==2.1.1
natsort==8.2.0
numpy==1.13.3
opencv_python_headless==4.9.0.80
OpenEXR==3.3.2
optuna==3.0.6
pandas==1.1.5
Pillow==11.1.0
scikit_image==0.17.2
scikit_learn==1.6.1
scipy==1.5.4
seaborn==0.11.2
skimage==0.0
torch==1.10.1+cu102
torchvision==0.11.2+cu102
tqdm==4.64.0
